refactor(flutter/a11y_assessments): move DynamicTitle to wrap useCase
Updates previous fix to wrap the DynamicTitle around the useCase as
recommended in PR review. WIP: New changes have caused the tests
checking for the unique page title on each use case page are now
showing empty labels. Committing to reference to reviewer.
