+++ dev/a11y_assessments/lib/use_cases/check_box_list_tile.dart
@@ -14,7 +14,7 @@ class CheckBoxListTile extends UseCase {
   String get route => '/check-box-list-tile';
 
   @override
-  Widget get mainWidget => _MainWidget();
+  Widget build(BuildContext context) => _MainWidget();
 }
 
 class _MainWidget extends StatefulWidget {
