+++ dev/a11y_assessments/lib/use_cases/date_picker.dart
@@ -14,7 +14,7 @@ class DatePickerUseCase extends UseCase {
   String get route => '/date-picker';
 
   @override
-  Widget get mainWidget => const _MainWidget();
+  Widget build(BuildContext context) => const _MainWidget();
 }
 
 class _MainWidget extends StatefulWidget {
