+++ dev/a11y_assessments/lib/use_cases/material_banner.dart
@@ -14,7 +14,7 @@ class MaterialBannerUseCase extends UseCase {
   String get route => '/material_banner';
 
   @override
-  Widget get mainWidget => const MainWidget();
+  Widget build(BuildContext context) => const MainWidget();
 }
 
 class MainWidget extends StatefulWidget {
