+++ dev/a11y_assessments/lib/use_cases/navigation_drawer.dart
@@ -33,7 +33,7 @@ class NavigationDrawerUseCase extends UseCase {
   String get route => '/navigation-drawer';
 
   @override
-  Widget get mainWidget => const NavigationDrawerExample();
+  Widget build(BuildContext context) => const NavigationDrawerExample();
 }
 
 class NavigationDrawerExample extends StatefulWidget {
