+++ dev/a11y_assessments/lib/use_cases/text_field.dart
@@ -14,7 +14,7 @@ class TextFieldUseCase extends UseCase {
   String get route => '/text-field';
 
   @override
-  Widget get mainWidget => _MainWidget();
+  Widget build(BuildContext context) => _MainWidget();
 }
 
 class _MainWidget extends StatelessWidget {
