+++ dev/a11y_assessments/lib/use_cases/text_field_password.dart
@@ -14,7 +14,7 @@ class TextFieldPasswordUseCase extends UseCase {
   String get route => '/text-field-password';
 
   @override
-  Widget get mainWidget => _MainWidget();
+  Widget build(BuildContext context) => _MainWidget();
 }
 
 class _MainWidget extends StatelessWidget {
